# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriCycle API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  farmer: "farmer",
  aggregator: "aggregator",
  factory: "factory",
} as const;

export interface RegisterRequest {
  name: string;
  phone: string;
  /** @minLength 6 */
  password: string;
  role: RegisterRequestRole;
  location: string;
}

export interface LoginRequest {
  phone: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  farmer: "farmer",
  aggregator: "aggregator",
  factory: "factory",
} as const;

export interface User {
  id: number;
  name: string;
  phone: string;
  role: UserRole;
  location: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type UpdateProfileRequestRole =
  (typeof UpdateProfileRequestRole)[keyof typeof UpdateProfileRequestRole];

export const UpdateProfileRequestRole = {
  farmer: "farmer",
  aggregator: "aggregator",
  factory: "factory",
} as const;

export interface UpdateProfileRequest {
  name?: string;
  role?: UpdateProfileRequestRole;
  location?: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}